% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_ops.R
\name{adjusted_ops}
\alias{adjusted_ops}
\title{Adjusted OPS}
\usage{
adjusted_ops(singles, doubles, triples, hr, bb, pa, ab, r, ...)
}
\arguments{
\item{singles}{# of singles}

\item{doubles}{# of doubles}

\item{triples}{# of triples}

\item{hr}{# of home runs}

\item{bb}{# of walks (user can add ibb and hbp to walks if desired)}

\item{pa}{# of plate appearances}

\item{ab}{# of at bats}

\item{r}{# of runs}

\item{...}{options}
}
\description{
A regression based ops model
}
\examples{
adjusted_ops()
}
\keyword{baseball}
\keyword{ops}

