% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_data.R
\name{baseball_data}
\alias{baseball_data}
\title{Get Data}
\usage{
baseball_data(data_set)
}
\arguments{
\item{file}{name of data set}
}
\description{
Get Data
}
\examples{
baseball_data(MLB)
}
\keyword{baseball}
\keyword{data}

