% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hit_summary}
\alias{hit_summary}
\title{Summary of Hitting}
\usage{
hit_summary(x = hitting)
}
\arguments{
\item{x}{The dataset to be summarize, by default hitting.}
}
\value{
A summary of the dataset hitting
}
\description{
Summary of Hitting
}
