% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{team_avg}
\alias{team_avg}
\title{Team Average Plot}
\usage{
team_avg(x = hitting, tm = "MIN")
}
\arguments{
\item{x}{The dataset to be summarize, by default hitting.}

\item{tm}{The team to be looked for their averages over time}
}
\value{
A plot of the averages of all teams, with a smoothing curve, and the team of interest highlighted in red points.
}
\description{
Team Average Plot
}
