% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_batting_2016.R
\name{team_batting_2016}
\alias{team_batting_2016}
\title{2016 Team Batting Stats}
\usage{
baseball_data(team_batting_2016)
}
\description{
2016 Team Batting Stats
}
\examples{
baseball_data(team_batting_2016)
}
\keyword{Batting}
\keyword{Team}

